# perf(clustering): store entity only once in lmdb