@bungle 2023-11-15
If I combine this to the `streaming branch` (that also removes LMDB drop) (https://github.com/Kong/kong/pull/11632) here are the results:

```
LMDB shared memory size: 216M

--------------------------------------------------------------------------------------------------
- Memory Usage (RSS) Report for Data Plane Worker Process:                                       -
--------------------------------------------------------------------------------------------------
Process Name                     PID    Max Mem  Min Mem  Mean Mem  Median Mem  Standard Deviation
nginx: worker process            46694     379M     313M      351M        354M                 18M
nginx: privileged agent process  46695     252M     245M      247M        246M                1.3M
                                           631M     558M      598M        600M                    
```

**465 MB DROP** in LMDB shared memory size (compared to master)
**800 MB DROP** in Privileged agent median memory usage (compared to master)
**1469 MB DROP** in total max memory usage