@chronolaw 2023-11-16
I read the code, I think that the core is `global_query_cache_key` (id with `*`), if a key contains `*`, then we will query twice to get the real value.

But I can not understand why it can reduce the memory usage.