@chronolaw 2023-11-16
Ah, I saw the only one writing action is `t:set(item_cache_key, item_marshalled)`, the others write the cache key.

But the format of cache key is still a little mysterious to me, especially `ws_id == "*"`.