@chronolaw 2023-11-16
In lua-resty-lmdb we introduce a new api `prefix()` (https://github.com/Kong/lua-resty-lmdb/pull/48), so the key may not be shorten, we will use prefix to query keys.