@bungle 2023-11-16
> But the format of cache key is still a little mysterious to me, especially `ws_id == "*"`.

Yes, this PR is not about to be merged, this is just a PoC and showcase that we can with simple changes get huge benefits on memory usage.

1. we need to get rid of lmdb drop (see: https://github.com/Kong/kong/pull/11632/commits/c90a5aae7c60f00d2d265c2791e732411ff30bc0)
2. we need better model for data in lmdb (this PR shows a simple change that already gives huge benefits)

So I feel we need to have data model in lmdb that works greatly for incremental changes (the commit above does it, but it is ugly), AND we need a data model that is efficient (this PR here tries to make it a bit more efficient, but there is much more we can do).