@bungle 2023-11-16
> In lua-resty-lmdb we introduce a new api `prefix()` ([Kong/lua-resty-lmdb#48](https://github.com/Kong/lua-resty-lmdb/pull/48)), so the key may not be shorten, we will use prefix to query keys.

@chronolaw, yes this is exactly what I am after in these conversations and discussions. Let's put our smart heads together and think. We already have evidence of what could work. But all the ideas are welcomed.