@bungle 2023-11-16
> In lua-resty-lmdb we introduce a new api prefix() (https://github.com/Kong/lua-resty-lmdb/pull/48), so the key may not be shorten, we will use prefix to query keys.

Checked that feature. Key names can still be shortened if we want, while maintaining some kind of prefixes still. This is a great feature.